Restructure all the structured data stuff
