--- src/flickypedia/apis/structured_data/__init__.py
@@ -1,25 +1,0 @@
-from .create_structured_data import (
-    create_copyright_status_statement,
-    create_date_taken_statement,
-    create_flickr_creator_statement,
-    create_flickr_photo_id_statement,
-    create_license_statement,
-    create_location_statement,
-    create_posted_to_flickr_statement,
-    create_sdc_claims_for_existing_flickr_photo,
-    create_sdc_claims_for_new_flickr_photo,
-    create_source_statement,
-)
-
-__all__ = [
-    "create_copyright_status_statement",
-    "create_date_taken_statement",
-    "create_flickr_creator_statement",
-    "create_flickr_photo_id_statement",
-    "create_license_statement",
-    "create_location_statement",
-    "create_posted_to_flickr_statement",
-    "create_sdc_claims_for_existing_flickr_photo",
-    "create_sdc_claims_for_new_flickr_photo",
-    "create_source_statement",
-]
