+++ src/flickypedia/photos.py
@@ -8,9 +8,11 @@ import typing
 from flask import current_app
 from flickr_photos_api import SinglePhoto, Size
 
-from flickypedia.apis.structured_data import create_sdc_claims_for_new_flickr_photo
 from flickypedia.duplicates import find_duplicates, DuplicateInfo
-from flickypedia.types.structured_data import NewClaims
+from flickypedia.structured_data import (
+    NewClaims,
+    create_sdc_claims_for_new_flickr_photo,
+)
 
 
 def size_at(sizes: list[Size], *, desired_size: str) -> Size:
