+++ src/flickypedia/structured_data/__init__.py
@@ -1,3 +1,39 @@
+from .claims import (
+    create_sdc_claims_for_existing_flickr_photo,
+    create_sdc_claims_for_new_flickr_photo,
+)
+from .types import (
+    DataValue,
+    ExistingClaims,
+    ExistingStatement,
+    NewClaims,
+    NewStatement,
+    Qualifiers,
+    Snak,
+    Value,
+    WikidataDatePrecision,
+    render_wikidata_date,
+    to_wikidata_entity_value,
+)
+from .wikidata_entities import WikidataEntities, get_wikidata_entity_label
 from .wikidata_properties import WikidataProperties, get_wikidata_property_label
 
-__all__ = ["WikidataProperties", "get_wikidata_property_label"]
+__all__ = [
+    "DataValue",
+    "ExistingClaims",
+    "ExistingStatement",
+    "NewClaims",
+    "NewStatement",
+    "Qualifiers",
+    "Snak",
+    "Value",
+    "WikidataDatePrecision",
+    "WikidataEntities",
+    "WikidataProperties",
+    "create_sdc_claims_for_existing_flickr_photo",
+    "create_sdc_claims_for_new_flickr_photo",
+    "get_wikidata_entity_label",
+    "get_wikidata_property_label",
+    "render_wikidata_date",
+    "to_wikidata_entity_value",
+]
