+++ src/flickypedia/uploadr/__init__.py
@@ -25,11 +25,11 @@ from flickypedia.apis.flickr import (
     create_bot_comment_text,
     create_default_user_comment_text,
 )
-from flickypedia.apis.structured_data.wikidata import (
-    get_entity_label,
+from flickypedia.structured_data import (
+    get_wikidata_entity_label,
+    get_wikidata_property_label,
     render_wikidata_date,
 )
-from flickypedia.structured_data import get_wikidata_property_label
 from .views import (
     about,
     bookmarklet,
@@ -129,7 +129,7 @@ def create_app(
     app.jinja_env.filters["bookmarklet"] = create_bookmarklet
 
     app.jinja_env.filters["wikidata_property_label"] = get_wikidata_property_label
-    app.jinja_env.filters["wikidata_entity_label"] = get_entity_label
+    app.jinja_env.filters["wikidata_entity_label"] = get_wikidata_entity_label
     app.jinja_env.filters["wikidata_date"] = render_wikidata_date
 
     app.jinja_env.filters["bot_comment_text"] = create_bot_comment_text
